#28353 update format to explicitly define storage root for each template
Template paths each have a storage root and a definition. Previously the template paths defined in the (single root) default config were in the format `template_name: 'template_path'` taking advantage of some "magic" where using this simplified format Toolkit would assume the storage `root_name` was "primary" and use `'template_path'` as the definition.

This was also handy due to the requirement that your primary stage root is named "primary". This requirement is in the process of going away so there should be no magic involved. We require explicit values for the template path `definition` and `root_name` keys so there is no question or magic. This also matches up with the format of the default multi-root config(s) and will make it much easier for studios to migrate to a multi-root setup in the future should they need to.