#44077 Initial release candidate for default2 config (#16)
These changes represent the full restructuring and organization of the default2 config. This includes a new environment structure that relies heavily on includes. All bundle configurations are isolated in single files, making it clear where to go to manually curate environments. 

In addition to the env changes, tk-multi-publish2 is now the default publish app. It is configured to use the publish hooks that now live with the engines and in the app itself. 

Softimage is no longer included in the default configuration